@@ -31,15 +31,23 @@ from .providers.orchestration.common import (
     EXTENSION_MONIKER_TO_ALIAS_MAP,
     TRUST_SETTING_KEYS,
     X509_ISSUER_REF_KEYS,
+    AuthenticationSaslType,
     CloneSummaryMode,
     CloneTemplateMode,
     CloneTemplateParams,
     ConfigSyncModeType,
+    DataflowEndpointAuthenticationType,
+    DataflowEndpointFabricPathType,
+    DataflowEndpointKafkaAcksType,
     IdentityUsageType,
+    KafkaCloudEventAttributeType,
+    KafkaCompressionType,
+    KafkaPartitionStrategyType,
     KubernetesDistroType,
     ListenerProtocol,
     MqMemoryProfile,
     MqServiceType,
+    MqttRetainType,
     SchemaFormat,
     SchemaType,
     TlsKeyAlgo,
@@ -141,6 +149,12 @@ def load_iotops_arguments(self, _):
             options_list=["--custom-role-id"],
             help="Fully qualified role definition Id in the following format: "
             "/subscriptions/{subscriptionId}/providers/Microsoft.Authorization/roleDefinitions/{roleId}",
+        )
+        context.argument(
+            "skip_role_assignments",
+            options_list=["--skip-ra"],
+            arg_type=get_three_state_flag(),
+            help="When used the role assignment step of the operation will be skipped.",
         )
 
     with self.argument_context("iot ops identity") as context:
@@ -286,22 +300,557 @@ def load_iotops_arguments(self, _):
             "profile_instances",
             type=int,
             options_list=["--profile-instances"],
-            help="The number of pods that run associated dataflows. Min value: 1, max value: 20."
+            help="The number of pods that run associated dataflows. Min value: 1, max value: 20.",
         )
         context.argument(
             "log_level",
             options_list=["--log-level"],
             help="The level of detail given in diagnostic logs. "
             "Levels: 'error', 'warn', 'info', 'debug', 'trace', 'off'. "
-            "For advanced usage you can specify a comma-separated list of module=level pairs. "
+            "For advanced usage you can specify a comma-separated list of module=level pairs.",
         )
 
     with self.argument_context("iot ops dataflow endpoint") as context:
+        # TODO: move dataflow params to separate file
         context.argument(
             "endpoint_name",
             options_list=["--name", "-n"],
             help="Dataflow endpoint name.",
         )
+        context.argument(
+            "authentication_type",
+            options_list=["--auth-type"],
+            choices=CaseInsensitiveList(
+                [
+                    DataflowEndpointAuthenticationType.ACCESSTOKEN.value,
+                    DataflowEndpointAuthenticationType.SASL.value,
+                    DataflowEndpointAuthenticationType.SERVICEACCESSTOKEN.value,
+                    DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                    DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                    DataflowEndpointAuthenticationType.X509.value,
+                ]
+            ),
+            help="The authentication type for the dataflow endpoint. "
+            "Note: When not specified, the authentication type is determinded by other authentication parameters.",
+        )
+        context.argument(
+            "client_id",
+            options_list=["--client-id", "--cid"],
+            help="The client ID of the user assigned identity",
+            arg_group="User Assigned Managed Identity",
+        )
+        context.argument(
+            "scope",
+            options_list=["--scope"],
+            help="Resource identifier (application ID URI) of the resource, affixed with the .default suffix.",
+            arg_group="User Assigned Managed Identity",
+        )
+        context.argument(
+            "tenant_id",
+            options_list=["--tenant-id", "--tid"],
+            help="The tenant ID of the user assigned identity.",
+            arg_group="User Assigned Managed Identity",
+        )
+        context.argument(
+            "no_auth",
+            options_list=["--no-auth"],
+            arg_type=get_three_state_flag(),
+            help="No authentication for the endpoint.",
+        )
+        context.argument(
+            "latency",
+            options_list=["--latency", "-l"],
+            help="The batching latency in seconds. Min value: 0, max value: 65535.",
+            type=int,
+            arg_group="Batching Configuration",
+        )
+        context.argument(
+            "message_count",
+            options_list=["--message-count", "--mc"],
+            help="Maximum number of messages in a batch. Min value: 0, max value: 4294967295.",
+            type=int,
+            arg_group="Batching Configuration",
+        )
+        context.argument(
+            "audience",
+            options_list=["--audience", "--aud"],
+            help="Audience of the service to authenticate against.",
+            arg_group="System Assigned Managed Identity",
+        )
+        context.argument(
+            "acks",
+            options_list=["--acks"],
+            arg_type=get_enum_type(DataflowEndpointKafkaAcksType, default=DataflowEndpointKafkaAcksType.ALL.value),
+            help="Level of acknowledgment from the Kafka broker to "
+            "ensure that the message sent by producer is successfully "
+            "written to the topic and replicated across the Kafka cluster.",
+        )
+        context.argument(
+            "batching_disabled",
+            options_list=["--disable-batching", "--db"],
+            arg_type=get_three_state_flag(),
+            help="Disable batching.",
+            arg_group="Batching Configuration",
+        )
+        context.argument(
+            "max_bytes",
+            options_list=["--max-bytes", "--mb"],
+            help="Maximum number of bytes in a batch.",
+            type=int,
+            arg_group="Batching Configuration",
+        )
+        context.argument(
+            "cloud_event_attribute",
+            options_list=["--cloud-event-attribute", "--cea"],
+            arg_type=get_enum_type(
+                KafkaCloudEventAttributeType,
+                default=KafkaCloudEventAttributeType.PROPAGATE.value,
+            ),
+            help="CloudEvent settings type to map events to cloud. "
+            "Different message format are required by different setting.",
+            arg_group="Cloud Event",
+        )
+        context.argument(
+            "compression",
+            options_list=["--compression"],
+            arg_type=get_enum_type(KafkaCompressionType, default=KafkaCompressionType.NONE.value),
+            help="Compression type for the messages sent to Kafka topics.",
+        )
+        context.argument(
+            "config_map_reference",
+            options_list=["--config-map-ref", "--cm"],
+            help="Config map reference for Trusted CA certificate for Kafka/MQTT endpoint. "
+            "Note: This ConfigMap should contain the CA certificate in PEM format. "
+            "The ConfigMap must be in the same namespace as the Kafka/MQTT data flow resource.",
+            arg_group="Transport Layer Security (TLS)",
+        )
+        context.argument(
+            "copy_broker_props_disabled",
+            options_list=["--disable-broker-props-copy", "--dbpc"],
+            arg_type=get_three_state_flag(),
+            help="Disable MQTT broker properties copy to Kafka user headers.",
+        )
+        context.argument(
+            "group_id",
+            options_list=["--group-id", "--gid"],
+            help="ID of consumer group that the data flow uses to read messages " "from the Kafka topic.",
+        )
+        context.argument(
+            "partition_strategy",
+            options_list=["--partition-strategy", "--ps"],
+            arg_type=get_enum_type(KafkaPartitionStrategyType, default=KafkaPartitionStrategyType.DEFAULT.value),
+            help="The partition handling strategy controls how messages are "
+            "assigned to Kafka partitions when sending them to Kafka topics.",
+        )
+        context.argument(
+            "sasl_type",
+            options_list=["--sasl-type"],
+            arg_type=get_enum_type(AuthenticationSaslType, default=None),
+            help="The type of SASL authentication.",
+            arg_group="SASL Authentication",
+        )
+        context.argument(
+            "tls_disabled",
+            options_list=["--disable-tls"],
+            arg_type=get_three_state_flag(),
+            help="The data flow uses an insecure connection to the Kafka/MQTT broker.",
+            arg_group="Transport Layer Security (TLS)",
+        )
+        context.argument(
+            "client_id_prefix",
+            options_list=["--client-id-prefix"],
+            help="The client id prefix for MQTT client. "
+            "Note: Changing the client ID prefix after IoT "
+            "Operations deployment might result in data loss.",
+        )
+        context.argument(
+            "keep_alive",
+            options_list=["--keep-alive", "--ka"],
+            help="The maximum time in seconds that the data flow client can "
+            "be idle before sending a PINGREQ message to the broker. "
+            "Min value: 0.",
+            type=int,
+        )
+        context.argument(
+            "max_inflight_messages",
+            options_list=["--max-inflight-msg", "--mim"],
+            help="The maximum number of inflight messages that the data flow MQTT client can have. Min value: 0.",
+            type=int,
+        )
+        context.argument(
+            "protocol",
+            options_list=["--protocol"],
+            arg_type=get_enum_type(ListenerProtocol, default=ListenerProtocol.MQTT.value),
+            help="Protocol to use for client connections.",
+        )
+        context.argument(
+            "qos",
+            options_list=["--qos"],
+            help="Quality of Service (QoS) level for the MQTT messages. Only 0 or 1 are supported.",
+            type=int,
+        )
+        context.argument(
+            "retain",
+            options_list=["--retain"],
+            arg_type=get_enum_type(MqttRetainType, default=MqttRetainType.KEEP.value),
+            help="Retain setting to specify whether the data flow should keep the retain "
+            "flag on MQTT messages. Setting this ensures whether or not the remote broker "
+            "has the same messages retained as the local broker.",
+        )
+        context.argument(
+            "session_expiry",
+            options_list=["--session-expiry"],
+            help="The session expiry interval in seconds for the data flow MQTT client. " "Min value: 0.",
+            type=int,
+        )
+
+    for cmd_space in ["iot ops dataflow endpoint create adx", "iot ops dataflow endpoint update adx"]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "database_name",
+                options_list=["--database", "--db"],
+                help="The name of the Azure Data Explorer database.",
+            )
+            context.argument(
+                "host",
+                options_list=["--host"],
+                help="Host of the Azure Data Explorer is "
+                "Azure Data Explorer cluster URI. In the form "
+                "of https://cluster.region.kusto.windows.net",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create adls",
+        "iot ops dataflow endpoint update adls",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains SAS token.",
+                arg_group="Access Token",
+            )
+            context.argument(
+                "storage_account_name",
+                options_list=["--storage-account", "--sa"],
+                help="The name of Azure Data Lake Storage Gen2 account.",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.ACCESSTOKEN.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create fabric-onelake",
+        "iot ops dataflow endpoint update fabric-onelake",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "workspace_name",
+                options_list=["--workspace"],
+                help="The Microsoft Fabric workspace name. Note: The default 'my workspace' isn't supported.",
+            )
+            context.argument(
+                "lakehouse_name",
+                options_list=["--lakehouse"],
+                help="The Microsoft Fabric lakehouse name under provided workspace.",
+            )
+            context.argument(
+                "path_type",
+                options_list=["--path-type", "--pt"],
+                arg_type=get_enum_type(DataflowEndpointFabricPathType, default=None),
+                help="The type of path used in OneLake.",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create eventhub",
+        "iot ops dataflow endpoint update eventhub",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "eventhub_namespace",
+                options_list=["--eventhub-namespace", "--ehns"],
+                help="The name of the Event Hubs namespace.",
+            )
+            context.argument(
+                "latency",
+                options_list=["--latency", "-l"],
+                help="The batching latency in milliseconds. Min value: 0, max value: 65535.",
+                type=int,
+                arg_group="Batching Configuration",
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains event hub connection string. "
+                "Note: The secret must be in the same namespace as the Kafka data flow endpoint. "
+                "The secret must have both the username and password as key-value pairs. "
+                "For more information about secret format, please refer to link in command description.",
+                arg_group="SASL Authentication",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.SASL.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create fabric-realtime",
+        "iot ops dataflow endpoint update fabric-realtime",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "host",
+                options_list=["--host"],
+                help="Host of the Fabric real-time is the "
+                "'Bootstrap server' value. Can be found in "
+                "event stream destination -- 'SAS Key Authentication' "
+                "section. In the form of *.servicebus.windows.net:9093",
+            )
+            context.argument(
+                "latency",
+                options_list=["--latency", "-l"],
+                help="The batching latency in milliseconds. Min value: 0, max value: 65535.",
+                type=int,
+                arg_group="Batching Configuration",
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains Connection string-primary key value. "
+                "Can be found in event stream destination -- 'SAS Key Authentication' section. "
+                "Note: The secret must be in the same namespace as the Kafka data flow endpoint. "
+                "For more information about secret format, please refer to link in command description.",
+                arg_group="SASL Authentication",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.SASL.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create custom-kafka",
+        "iot ops dataflow endpoint update custom-kafka",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "hostname",
+                options_list=["--hostname"],
+                help="The hostname of the Kafka broker host setting.",
+            )
+            context.argument(
+                "port",
+                options_list=["--port"],
+                help="The port number of the Kafka broker host setting.",
+                type=int,
+            )
+            context.argument(
+                "latency",
+                options_list=["--latency", "-l"],
+                help="The batching latency in milliseconds. Min value: 0, max value: 65535.",
+                type=int,
+                arg_group="Batching Configuration",
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name of the Kubernetes secret that contains the SASL token.",
+                arg_group="SASL Authentication",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.SASL.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create local-storage",
+        "iot ops dataflow endpoint update local-storage",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "pvc_reference",
+                options_list=["--pvc-ref"],
+                help="The name of the PersistentVolumeClaim (PVC) to use for local storage. "
+                "Note: The PVC must be in the same namespace as the data flow endpoint.",
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create local-mqtt",
+        "iot ops dataflow endpoint update local-mqtt",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "hostname",
+                options_list=["--hostname"],
+                help="The hostname of the local MQTT broker.",
+            )
+            context.argument(
+                "port",
+                options_list=["--port"],
+                help="The port number of the local MQTT broker.",
+                type=int,
+            )
+            context.argument(
+                "audience",
+                options_list=["--audience", "--aud"],
+                help="The audience of the Kubernetes service account token (SAT).",
+                arg_group="Kubernetes Service Account Token",
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains the X509 client certificate, private key "
+                "corresponding to the client certificate, and intermediate certificates for the client certificate "
+                "chain. "
+                "Note: The certificate and private key must be in PEM format and not password protected.",
+                arg_group="X509 Authentication",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SERVICEACCESSTOKEN.value,
+                        DataflowEndpointAuthenticationType.X509.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create eventgrid",
+        "iot ops dataflow endpoint update eventgrid",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "hostname",
+                options_list=["--hostname"],
+                help="The hostname of the event grid namespace. Can"
+                " be found in 'Http hostname' property. In the form "
+                "of NAMESPACE.REGION-1.ts.eventgrid.azure.net",
+            )
+            context.argument(
+                "port",
+                options_list=["--port"],
+                help="The port number of the event grid namespace.",
+                type=int,
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains the X509 client certificate, private key "
+                "corresponding to the client certificate, and intermediate certificates for the client certificate "
+                "chain. "
+                "Note: The certificate and private key must be in PEM format and not password protected.",
+                arg_group="X509 Authentication",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.X509.value,
+                    ]
+                ),
+            )
+
+    for cmd_space in [
+        "iot ops dataflow endpoint create custom-mqtt",
+        "iot ops dataflow endpoint update custom-mqtt",
+    ]:
+        with self.argument_context(cmd_space) as context:
+            context.argument(
+                "hostname",
+                options_list=["--hostname"],
+                help="The hostname of the custom MQTT broker host setting.",
+            )
+            context.argument(
+                "port",
+                options_list=["--port"],
+                help="The port number of the custom MQTT broker host setting.",
+                type=int,
+            )
+            context.argument(
+                "secret_name",
+                options_list=["--secret-name", "-s"],
+                help="The name for the kubernetes secret that contains the X509 client certificate, private key "
+                "corresponding to the client certificate, and intermediate certificates for the client certificate "
+                "chain. "
+                "Note: The certificate and private key must be in PEM format and not password protected.",
+                arg_group="X509 Authentication",
+            )
+            context.argument(
+                "sami_audience",
+                options_list=["--sami-audience", "--sami-aud"],
+                help="The audience of the system assigned managed identity.",
+                arg_group="System Assigned Managed Identity",
+            )
+            context.argument(
+                "sat_audience",
+                options_list=["--sat-audience", "--sat-aud"],
+                help="The audience of the Kubernetes service account token (SAT).",
+                arg_group="Kubernetes Service Account Token",
+            )
+            context.argument(
+                "authentication_type",
+                options_list=["--auth-type"],
+                choices=CaseInsensitiveList(
+                    [
+                        DataflowEndpointAuthenticationType.SERVICEACCESSTOKEN.value,
+                        DataflowEndpointAuthenticationType.SYSTEMASSIGNED.value,
+                        DataflowEndpointAuthenticationType.USERASSIGNED.value,
+                        DataflowEndpointAuthenticationType.X509.value,
+                    ]
+                ),
+            )
 
     with self.argument_context("iot ops broker") as context:
         context.argument(
@@ -547,6 +1096,7 @@ def load_iotops_arguments(self, _):
                 "enable_rsync_rules",
                 options_list=["--enable-rsync"],
                 arg_type=get_three_state_flag(),
+                deprecate_info=context.deprecate(target="--enable-rsync", redirect='az iot ops rsync enable'),
                 help="Resource sync rules will be included in the IoT Operations deployment.",
             )
             context.argument(
@@ -555,6 +1105,15 @@ def load_iotops_arguments(self, _):
                 arg_type=get_three_state_flag(),
                 help="Ensure the latest IoT Ops CLI is being used, raising an error if an upgrade is available.",
             )
+            # Cluster Precheck opt-in
+            context.argument(
+                "check_cluster",
+                options_list=["--check-cluster"],
+                arg_type=get_three_state_flag(),
+                help="Enforce a check for minimum cluster requirements before bootstrapping.",
+                is_preview=True,
+            )
+
             # Schema Registry
             context.argument(
                 "schema_registry_resource_id",
@@ -798,12 +1357,6 @@ def load_iotops_arguments(self, _):
             options_list=["--spc"],
             help="The default secret provider class name for secret sync enablement. "
             "The default pattern is 'spc-ops-{hash}'.",
-        )
-        context.argument(
-            "skip_role_assignments",
-            options_list=["--skip-ra"],
-            arg_type=get_three_state_flag(),
-            help="When used the role assignment step of the operation will be skipped.",
         )
         context.argument(
             "instance_name",
@@ -1107,4 +1660,43 @@ def load_iotops_arguments(self, _):
             help="Use the self-hosted oidc issuer for federation. Only applicable if "
             "user-assigned managed identities are associated to the model instance.",
             arg_group="Cluster Target",
+        )
+
+    with self.argument_context("iot ops rsync") as context:
+        context.argument(
+            "instance_name",
+            options_list=["--instance", "-i", "-n"],
+            help="IoT Operations instance name.",
+        )
+        context.argument(
+            "rule_ops_name",
+            options_list=["--rule-ops-name"],
+            help="The desired name of the resource sync rule to create for IoT Operations.",
+            arg_group="Custom Rule",
+        )
+        context.argument(
+            "rule_adr_name",
+            options_list=["--rule-adr-name"],
+            help="The desired name of the resource sync rule to create for Device Registry.",
+            arg_group="Custom Rule",
+        )
+        context.argument(
+            "rule_ops_pri",
+            type=int,
+            options_list=["--rule-ops-pri"],
+            help="The desired priority of the resource sync rule to create for IoT Operations.",
+            arg_group="Custom Rule",
+        )
+        context.argument(
+            "rule_adr_pri",
+            type=int,
+            options_list=["--rule-adr-pri"],
+            help="The desired priority of the resource sync rule to create for Device Registry.",
+            arg_group="Custom Rule",
+        )
+        context.argument(
+            "k8_bridge_sp_oid",
+            options_list=["--k8-bridge-sp-oid"],
+            help="Provide an explicit K8 Bridge service principal OID for the custom location role assignment. "
+            "This is useful if the logged-in principal does not have permission to query apps from MS Graph.",
         )
