@@ -4,31 +4,36 @@
 # Licensed under the MIT License. See License file in the project root for license information.
 # ----------------------------------------------------------------------------------------------
 
+from typing import Any, Dict, Optional
+
 from knack.log import get_logger
 from kubernetes.client.exceptions import ApiException
 from kubernetes.client.models import V1Node, V1NodeList
 from rich.padding import Padding
 from rich.table import Table
-from typing import Any, Dict
 
-from .check_manager import CheckManager
-from .user_strings import NO_NODES_MSG, UNABLE_TO_FETCH_NODES_MSG
+from ....common import CheckTaskStatus
 from ..common import (
+    ACSA_MIN_NODE_KERNEL_VERSION,
     AIO_SUPPORTED_ARCHITECTURES,
     COLOR_STR_FORMAT,
     DISPLAY_BYTES_PER_GIGABYTE,
     MIN_NODE_MEMORY,
     MIN_NODE_STORAGE,
-    MIN_NODE_VCPU
+    MIN_NODE_VCPU,
+    NODE_CONTROL_PLANE_LABEL,
 )
-from ....common import CheckTaskStatus
-
+from .check_manager import CheckManager
+from .user_strings import NO_NODES_MSG, UNABLE_TO_FETCH_NODES_MSG
 
 logger = get_logger(__name__)
 
 
-def check_nodes(as_list: bool = False) -> Dict[str, Any]:
+def check_nodes(
+    as_list: bool = False, kernel_version_check: Optional[bool] = False, storage_space_check: Optional[bool] = True
+) -> Dict[str, Any]:
     from ...base import client
+
     check_manager = CheckManager(check_name="evalClusterNodes", check_desc="Evaluate cluster nodes")
     padding = (0, 0, 0, 8)
     target = "cluster/nodes"
@@ -52,14 +57,19 @@ def check_nodes(as_list: bool = False) -> Dict[str, Any]:
     else:
         if not nodes or not nodes.items:
             target_display = Padding(NO_NODES_MSG, padding)
-            check_manager.add_target_eval(
-                target_name=target, status=CheckTaskStatus.error.value, value=NO_NODES_MSG
-            )
+            check_manager.add_target_eval(target_name=target, status=CheckTaskStatus.error.value, value=NO_NODES_MSG)
             check_manager.add_display(target_name=target, display=target_display)
             return check_manager.as_dict()
 
-        check_manager.add_target_eval(target_name=target, status=CheckTaskStatus.success.value, value={"len(cluster/nodes)": len(nodes.items)})
-        table = _generate_node_table(check_manager, nodes)
+        check_manager.add_target_eval(
+            target_name=target, status=CheckTaskStatus.success.value, value={"len(cluster/nodes)": len(nodes.items)}
+        )
+        table = _generate_node_table(
+            check_manager=check_manager,
+            nodes=nodes,
+            acs_kernel_check=kernel_version_check,
+            storage_space_check=storage_space_check,
+        )
 
         check_manager.add_display(target_name=target, display=Padding("Node Resources", padding))
         check_manager.add_display(target_name=target, display=Padding(table, padding))
@@ -67,30 +77,50 @@ def check_nodes(as_list: bool = False) -> Dict[str, Any]:
     return check_manager.as_dict(as_list)
 
 
-def _generate_node_table(check_manager: CheckManager, nodes: V1NodeList) -> Table:
+def _generate_node_table(
+    check_manager: CheckManager,
+    nodes: V1NodeList,
+    acs_kernel_check: Optional[bool] = False,
+    storage_space_check: Optional[bool] = False,
+) -> Table:
     from kubernetes.utils import parse_quantity
+
+    from ....util.machinery import scoped_semver_import
+
+    semver = scoped_semver_import()
+
     # prep table
-    table = Table(
-        show_header=True, header_style="bold", show_lines=True, caption_justify="left"
-    )
+    table = Table(show_header=True, header_style="bold", show_lines=True, caption_justify="left")
     for column_name, justify in [
         ("Name", "left"),
-        ("Architecture", "right"),
-        ("CPU (vCPU)", "right"),
-        ("Memory (GB)", "right"),
-        ("Storage (GB)", "right"),
+        ("Architecture", "left"),
+        *([("Kernel version", "left")] if acs_kernel_check else []),
+        ("CPU (vCPU)", "left"),
+        ("Memory (GB)", "left"),
+        *([("Ephemeral\nStorage (GB)", "left")] if storage_space_check else []),
     ]:
         table.add_column(column_name, justify=f"{justify}")
-    table.add_row(*[COLOR_STR_FORMAT.format(color="cyan", value=value) for value in [
-        "Minimum requirements",
-        ", ".join(AIO_SUPPORTED_ARCHITECTURES),
-        MIN_NODE_VCPU,
-        MIN_NODE_MEMORY[:-1],
-        MIN_NODE_STORAGE[:-1]
-    ]])
+    table.add_row(
+        *[
+            COLOR_STR_FORMAT.format(color="cyan", value=value)
+            for value in [
+                "Minimum requirements",
+                ", ".join(AIO_SUPPORTED_ARCHITECTURES),
+                *([ACSA_MIN_NODE_KERNEL_VERSION] if acs_kernel_check else []),
+                MIN_NODE_VCPU,
+                MIN_NODE_MEMORY[:-1],
+                *([MIN_NODE_STORAGE[:-1]] if storage_space_check else []),
+            ]
+        ]
+    )
+    single_node = len(nodes.items) == 1
     node: V1Node
     for node in nodes.items:
         node_name = node.metadata.name
+
+        # skip control plane node checks if not single node
+        is_control_plane_node = node.metadata.labels and NODE_CONTROL_PLANE_LABEL in node.metadata.labels
+        control_plane_only_node = is_control_plane_node and not single_node
 
         # check_manager target for node
         node_target = f"cluster/nodes/{node_name}"
@@ -100,50 +130,92 @@ def _generate_node_table(check_manager: CheckManager, nodes: V1NodeList) -> Tabl
         # build node table row
         row_status = CheckTaskStatus.success
         row_cells = []
-        for condition, expected, actual in [
+        table_tuples = [
             (
                 "info.architecture",
                 AIO_SUPPORTED_ARCHITECTURES,
                 node.status.node_info.architecture,
             ),
+            # Optional kernel version check
+            *(
+                [
+                    (
+                        "info.kernel_version",
+                        ACSA_MIN_NODE_KERNEL_VERSION,
+                        node.status.node_info.kernel_version,
+                    )
+                ]
+                if acs_kernel_check
+                else []
+            ),
             (
-                "condition.cpu",
+                "allocatable.cpu",
                 MIN_NODE_VCPU,
-                parse_quantity(node.status.capacity.get("cpu", 0)),
+                parse_quantity(node.status.allocatable.get("cpu", 0)),
             ),
             (
-                "condition.memory",
+                "allocatable.memory",
                 MIN_NODE_MEMORY,
-                parse_quantity(node.status.capacity.get("memory", 0)),
+                parse_quantity(node.status.allocatable.get("memory", 0)),
             ),
-            (
-                "condition.ephemeral-storage",
-                MIN_NODE_STORAGE,
-                parse_quantity(node.status.capacity.get("ephemeral-storage", 0)),
+            # Optional storage space check
+            *(
+                [
+                    (
+                        "allocatable.ephemeral-storage",
+                        MIN_NODE_STORAGE,
+                        parse_quantity(node.status.allocatable.get("ephemeral-storage", 0)),
+                    )
+                ]
+                if storage_space_check
+                else []
             ),
-        ]:
+        ]
+        for condition, expected, actual in table_tuples:
             # determine strings, expected, status
             condition_str = f"{condition}>={expected}"
             displayed = actual
             cell_status = CheckTaskStatus.success
-            if isinstance(expected, list):
+            if condition == "info.architecture":
                 condition_str = f"{condition} in ({','.join(expected)})"
                 if actual not in expected:
-                    row_status = CheckTaskStatus.error
                     cell_status = CheckTaskStatus.error
+            elif condition == "info.kernel_version":
+                # Only check major and minor versions
+                truncated_actual = ".".join(actual.split(".")[:2])
+                parsed_actual = semver.parse(truncated_actual, optional_minor_and_patch=True)
+                parsed_expected = semver.parse(expected, optional_minor_and_patch=True)
+
+                if parsed_actual < parsed_expected:
+                    cell_status = CheckTaskStatus.error
+
             else:
                 displayed = _get_display_number(displayed, expected)
                 expected = parse_quantity(expected)
                 if actual < expected:
-                    row_status = CheckTaskStatus.error
                     cell_status = CheckTaskStatus.error
                 actual = int(actual)
+                # display as "xxG"
+                if condition in ["allocatable.memory", "allocatable.ephemeral-storage"]:
+                    actual = f"{int(actual / DISPLAY_BYTES_PER_GIGABYTE)}G"
+
+            # if control plane only node, mark as skipped
+            if control_plane_only_node:
+                cell_status = CheckTaskStatus.skipped
+                displayed = f"[dim]{displayed}[/dim]"
+
+            # row status is error if any cell is error
+            if cell_status == CheckTaskStatus.error:
+                row_status = CheckTaskStatus.error
 
             check_manager.add_target_conditions(target_name=node_target, conditions=[condition_str])
             check_manager.add_target_eval(target_name=node_target, status=cell_status.value, value={condition: actual})
             row_cells.append(COLOR_STR_FORMAT.format(color=cell_status.color, value=displayed))
 
-        # overall node name color
+        # skip whole node table row if control_plane_only_node
+        if control_plane_only_node:
+            row_status = CheckTaskStatus.skipped
+            node_name = f"[dim]{node_name}[/dim]"
         table.add_row(COLOR_STR_FORMAT.format(color=row_status.color, value=node_name), *row_cells)
     return table
 
