@@ -29,14 +29,6 @@ def mocked_permission_manager(mocker):
     yield patched
 
 
-@pytest.fixture
-def mocked_verify_custom_locations_enabled(mocker):
-    patched = mocker.patch(
-        "azext_edge.edge.providers.orchestration.base.verify_custom_locations_enabled", autospec=True
-    )
-    yield patched
-
-
 ##
 
 
@@ -82,7 +74,7 @@ def mock_broker_config():
 def mocked_sleep(mocker):
     patched = {
         "az_client.sleep": mocker.patch("azext_edge.edge.util.az_client.sleep", autospec=True),
-        "work.sleep": mocker.patch("azext_edge.edge.providers.orchestration.work.sleep", autospec=True)
+        "work.sleep": mocker.patch("azext_edge.edge.providers.orchestration.work.sleep", autospec=True),
     }
     yield patched
 
@@ -98,6 +90,28 @@ def spy_work_displays(mocker):
 
 
 @pytest.fixture
+def mock_prechecks(mocker):
+    from azext_edge.edge.providers.check.base import deployment
+
+    yield {
+        "validate_cluster_prechecks": mocker.spy(deployment, "check_pre_deployment"),
+        "check_k8s_version": mocker.patch(
+            "azext_edge.edge.providers.check.base.deployment._check_k8s_version", autospec=True
+        ),
+        "check_storage_classes": mocker.patch(
+            "azext_edge.edge.providers.check.base.deployment._check_storage_classes", autospec=True
+        ),
+        "check_nodes": mocker.patch("azext_edge.edge.providers.check.base.deployment.check_nodes", autospec=True),
+    }
+
+
+@pytest.fixture
 def mocked_logger(mocker):
     patched = mocker.patch("azext_edge.edge.providers.orchestration.work.logger", autospec=True)
+    yield patched
+
+
+@pytest.fixture
+def mocked_verify_arc_cluster_config(mocker):
+    patched = mocker.patch("azext_edge.edge.providers.orchestration.work.verify_arc_cluster_config", autospec=True)
     yield patched
