@@ -36,8 +36,8 @@ from azext_edge.tests.generators import generate_random_string
 # TODO: Resturcture parameters into dict
 @pytest.mark.parametrize(
     "expected_resources_map, client_app_spc, client_app_secretsync,"
-    "public_file_name, private_file_name, expected_secret_sync,"
-    "expected_extension_properties",
+    "public_file_name, private_file_name, mocked_public_cert, mocked_private_cert,"
+    "expected_secret_sync, expected_extension_properties",
     [
         (
             {
@@ -56,6 +56,8 @@ from azext_edge.tests.generators import generate_random_string
             ),
             "/fake/path/certificate.der",
             "/fake/path/certificate.pem",
+            [build_mock_cert()],
+            [build_mock_cert()],
             get_mock_secretsync_record(
                 secretsync_name=OPCUA_CLIENT_CERT_SECRET_SYNC_NAME,
                 resource_group_name="mock-rg",
@@ -94,6 +96,8 @@ from azext_edge.tests.generators import generate_random_string
             ),
             "/fake/path/certificate.der",
             "/fake/path/certificate.pem",
+            [build_mock_cert()],
+            [build_mock_cert()],
             get_mock_secretsync_record(
                 secretsync_name=OPCUA_CLIENT_CERT_SECRET_SYNC_NAME,
                 resource_group_name="mock-rg",
@@ -131,6 +135,8 @@ from azext_edge.tests.generators import generate_random_string
             ),
             "/fake/path/certificate.der",
             "/fake/path/certificate.pem",
+            [build_mock_cert()],
+            [build_mock_cert()],
             get_mock_secretsync_record(
                 secretsync_name=OPCUA_CLIENT_CERT_SECRET_SYNC_NAME,
                 resource_group_name="mock-rg",
@@ -149,9 +155,10 @@ def test_client_add(
     mocker,
     mocked_cmd,
     mocked_read_file_content: Mock,
-    mocked_load_x509_cert: Mock,
     mocked_sleep: Mock,
     mocked_logger: Mock,
+    mocked_public_cert: dict,
+    mocked_private_cert: dict,
     expected_resources_map: dict,
     client_app_spc: dict,
     client_app_secretsync: dict,
@@ -160,6 +167,7 @@ def test_client_add(
     expected_secret_sync: dict,
     expected_extension_properties: dict,
     mocked_get_resource_client: Mock,
+    mocked_decode_certificate: Mock,
     mocked_instance: Mock,
     mocked_responses: responses,
 ):
@@ -175,6 +183,7 @@ def test_client_add(
     mocked_instance.find_existing_resources.return_value = expected_resources_map["resources"]
     mocked_get_resource_client().resources.get_by_id.return_value = {"id": "mock-id"}
     mocked_read_file_content.return_value = file_content
+    mocked_decode_certificate.side_effect = [mocked_public_cert, mocked_private_cert]
 
     if expected_resources_map["resources"]:
         # get secrets
@@ -274,8 +283,8 @@ def test_client_add(
 
 @pytest.mark.parametrize(
     "expected_resources_map, client_app_spc, client_app_secretsync,"
-    "public_file_name, private_file_name, subject_name, uri, cert_subject_name,"
-    "cert_uri, expected_error_type, expected_error_text",
+    "public_file_name, private_file_name, subject_name, uri,"
+    "mocked_public_cert, mocked_private_cert, expected_error_type, expected_error_text",
     [
         # no default spc
         (
@@ -289,8 +298,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             "subjectname",
             "uri",
-            "subjectname",
-            "uri",
+            [build_mock_cert()],
+            [build_mock_cert()],
             ResourceNotFoundError,
             "Please enable secret sync before adding certificate.",
         ),
@@ -314,8 +323,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             "subjectname",
             "uri",
-            "subjectname",
-            "uri",
+            [build_mock_cert()],
+            [build_mock_cert()],
             ResourceNotFoundError,
             "IoT Operations extension not found.",
         ),
@@ -331,8 +340,8 @@ def test_client_add(
             "/fake/path/prikey.pem",
             "subjectname",
             "uri",
-            "subjectname",
-            "uri",
+            [build_mock_cert()],
+            [build_mock_cert()],
             InvalidArgumentValueError,
             "Public key file name pubkey and private key file name prikey must match.",
         ),
@@ -348,8 +357,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             " ",
             "uri",
-            " ",
-            "uri",
+            [build_mock_cert(subject_name=" ", uri="uri")],
+            [build_mock_cert(subject_name=" ", uri="uri")],
             InvalidArgumentValueError,
             "Not able to extract subject name from the certificate. "
             "Please provide the correct subject name in certificate via --public-key-file.",
@@ -366,8 +375,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             "subjectname",
             " ",
-            "subjectname",
-            " ",
+            [build_mock_cert(subject_name="subjectname", uri=" ")],
+            [build_mock_cert(subject_name="subjectname", uri=" ")],
             InvalidArgumentValueError,
             "Not able to extract application URI from the certificate. "
             "Please provide the correct application URI in certificate via --public-key-file.",
@@ -384,8 +393,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             "subjectnamenotmatch",
             "uri",
-            "subjectname",
-            "uri",
+            [build_mock_cert(subject_name="subjectname", uri="uri")],
+            [build_mock_cert()],
             InvalidArgumentValueError,
             "Given --subject-name subjectnamenotmatch does not match certificate subject name subjectname. "
             "Please provide the correct subject name via --subject-name or correct certificate using --public-key-file."
@@ -402,8 +411,8 @@ def test_client_add(
             "/fake/path/certificate.pem",
             "subjectname",
             "urinotmatch",
-            "subjectname",
-            "uri",
+            [build_mock_cert()],
+            [build_mock_cert()],
             InvalidArgumentValueError,
             "Given --application-uri urinotmatch does not match certificate application URI uri. "
             "Please provide the correct application URI via --application-uri or correct certificate "
@@ -415,7 +424,7 @@ def test_client_add_errors(
     mocker,
     mocked_cmd,
     mocked_read_file_content: Mock,
-    mocked_load_x509_cert: Mock,
+    mocked_decode_certificate: Mock,
     mocked_sleep: Mock,
     expected_resources_map: dict,
     client_app_spc: dict,
@@ -424,8 +433,8 @@ def test_client_add_errors(
     private_file_name: str,
     subject_name: str,
     uri: str,
-    cert_subject_name: str,
-    cert_uri: str,
+    mocked_public_cert: list,
+    mocked_private_cert: list,
     expected_error_text: str,
     expected_error_type: Exception,
     mocked_get_resource_client: Mock,
@@ -444,7 +453,7 @@ def test_client_add_errors(
     mocked_instance.find_existing_resources.return_value = expected_resources_map["resources"]
     mocked_get_resource_client().resources.get_by_id.return_value = {"id": "mock-id"}
     mocked_read_file_content.return_value = file_content
-    mocked_load_x509_cert.return_value = build_mock_cert(subject_name=cert_subject_name, uri=cert_uri)
+    mocked_decode_certificate.side_effect = [mocked_public_cert, mocked_private_cert]
 
     if client_app_spc:
         # get secrets
