@@ -5,63 +5,24 @@
 # ----------------------------------------------------------------------------------------------
 
 import json
-from typing import Optional
 from unittest.mock import Mock
 
 import pytest
 import responses
 
-from azext_edge.edge.providers.orchestration.common import DATAFLOW_ENDPOINT_TYPE_SETTINGS
 from azext_edge.edge.commands_dataflow import (
     apply_dataflow_endpoint,
     delete_dataflow_endpoint,
     show_dataflow_endpoint,
     list_dataflow_endpoints,
 )
-from azext_edge.tests.edge.orchestration.resources.test_instances_unit import (
+from .conftest import get_dataflow_endpoint_endpoint, get_mock_dataflow_endpoint_record
+from ..test_instances_unit import (
     get_instance_endpoint,
     get_mock_instance_record,
 )
 
-from ....generators import generate_random_string
-from .conftest import get_base_endpoint, get_mock_resource
-
-
-def get_dataflow_endpoint_endpoint(
-    instance_name: str, resource_group_name: str, dataflow_endpoint_name: Optional[str] = None
-) -> str:
-    resource_path = f"/instances/{instance_name}/dataflowEndpoints"
-    if dataflow_endpoint_name:
-        resource_path += f"/{dataflow_endpoint_name}"
-    return get_base_endpoint(resource_group_name=resource_group_name, resource_path=resource_path)
-
-
-def get_mock_dataflow_endpoint_record(
-    dataflow_endpoint_name: str,
-    instance_name: str,
-    resource_group_name: str,
-    dataflow_endpoint_type: Optional[str] = None,
-    host: Optional[str] = None,
-    group_id: Optional[str] = None,
-) -> dict:
-    return get_mock_resource(
-        name=dataflow_endpoint_name,
-        resource_path=f"/instances/{instance_name}/dataflowEndpoints/{dataflow_endpoint_name}",
-        properties={
-            "authentication": {"method": "AccessToken"},
-            "accessTokenSecretRef": "mysecret",
-            "endpointType": dataflow_endpoint_type or "Kafka",
-            DATAFLOW_ENDPOINT_TYPE_SETTINGS[dataflow_endpoint_type or "Kafka"]: {
-                "tls": {"mode": "Enabled", "trustedCaCertificateConfigMapRef": "myconfigmap"},
-                "host": host or "myhost",
-                "consumerGroupId": group_id or "",
-            },
-            "provisioningState": "Succeeded",
-        },
-        resource_group_name=resource_group_name,
-        qualified_type="microsoft.iotoperations/instances/dataflowendpoints",
-        is_proxy_resource=True,
-    )
+from .....generators import generate_random_string
 
 
 def test_dataflow_endpoint_show(mocked_cmd, mocked_responses: responses):
